services/identity-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
confluent-kafka==2.3.0
pydantic-settings==2.0.3
structlog==23.2.0
opentelemetry-sdk==1.21.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0
httpx==0.25.2
python-multipart==0.0.6
email-validator==2.1.0
slowapi==0.1.9
